+++ src/interp/strategy.rs
@@ -1,6 +1,4 @@
 use super::*;
-use itertools::Itertools;
-use std::fmt;
 
 #[derive(Clone, Debug)]
 pub struct Branch {
@@ -19,7 +17,7 @@ impl Branch {
                 .iter()
                 .map(|par| Term::Var(par.tag_ctx(ctx_cnt)))
                 .collect(),
-            info: CallInfo::new(),
+            history: History::new(),
         };
 
         Branch {
@@ -35,7 +33,7 @@ impl Branch {
 
     pub fn clear_history(&mut self) {
         for call in self.calls.iter_mut() {
-            call.info.history.clear();
+            call.history.clear();
         }
     }
 
@@ -78,7 +76,7 @@ impl Branch {
         let idx = self
             .calls
             .iter()
-            .position(|call| call.info.history.naive_strategy_pred(n));
+            .position(|call| call.history.naive_strategy_pred(n));
 
         if let Some(idx) = idx {
             idx
@@ -93,8 +91,7 @@ impl Branch {
         assert!(!self.calls.is_empty());
 
         let idx = self.calls.iter().position(|call| {
-            call.info
-                .history
+            call.history
                 .struct_recur_strategy_pred(call.pred, &call.args)
         });
 
@@ -105,22 +102,6 @@ impl Branch {
             self.struct_recur_strategy()
         }
     }
-
-    #[allow(unused)]
-    pub fn conflit_driven_strategy(&mut self, cache: &mut ConflitCache) -> usize {
-        assert!(!self.calls.is_empty());
-
-        let idx = self
-            .calls
-            .iter()
-            .position(|call| cache.lookup(&call.info.path));
-
-        if let Some(idx) = idx {
-            idx
-        } else {
-            self.random_strategy()
-        }
-    }
 }
 
 #[derive(Clone, Debug)]
@@ -129,69 +110,7 @@ pub struct PredCall {
     pub pred: Ident,
     pub polys: Vec<TermType>,
     pub args: Vec<TermVal<IdentCtx>>,
-    pub info: CallInfo,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-struct PathNode {
-    rule_idx: usize,
-    call_idx: usize,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Path(Vec<PathNode>);
-
-impl Path {
-    pub fn new() -> Path {
-        Path(Vec::new())
-    }
-
-    pub fn clear(&mut self) {
-        self.0.clear();
-    }
-
-    pub fn push(&mut self, rule_idx: usize, call_idx: usize) {
-        self.0.push(PathNode { rule_idx, call_idx });
-    }
-}
-
-impl PartialOrd for Path {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        match self.0.len().cmp(&other.0.len()) {
-            std::cmp::Ordering::Less => {
-                for (node1, node2) in self.0.iter().zip(other.0.iter()) {
-                    if node1 != node2 {
-                        return None;
-                    }
-                }
-                Some(std::cmp::Ordering::Less)
-            }
-            std::cmp::Ordering::Equal => {
-                if self.0 == other.0 {
-                    Some(std::cmp::Ordering::Equal)
-                } else {
-                    None
-                }
-            }
-            std::cmp::Ordering::Greater => {
-                for (node1, node2) in self.0.iter().zip(other.0.iter()) {
-                    if node1 != node2 {
-                        return None;
-                    }
-                }
-                Some(std::cmp::Ordering::Greater)
-            }
-        }
-    }
-}
-
-impl std::fmt::Display for Path {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let nodes = self.0.iter().format_with(", ", |node, f| {
-            f(&format_args!("({}, {})", node.rule_idx, node.call_idx))
-        });
-        write!(f, "[{}]", nodes)
-    }
+    pub history: History,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
@@ -247,66 +166,7 @@ impl History {
     }
 }
 
-#[derive(Clone, Debug)]
-pub struct CallInfo {
-    // these two vector should have the same length, they are seperated for convenience.
-    pub history: History,
-    pub path: Path,
-}
-
-impl CallInfo {
-    pub fn new() -> CallInfo {
-        CallInfo {
-            history: History::new(),
-            path: Path::new(),
-        }
-    }
-}
-
-pub struct ConflitCache {
-    max_size: usize,
-    cache: Vec<Path>,
-    cursor: usize,
-}
-
-impl ConflitCache {
-    pub fn new(max_size: usize) -> ConflitCache {
-        ConflitCache {
-            max_size,
-            cache: std::iter::repeat_n(Path::new(), max_size).collect(),
-            cursor: 0,
-        }
-    }
-
-    pub fn lookup(&self, path: &Path) -> bool {
-        // println!("check: {}", path);
-        // for path in self.cache.iter() {
-        //     println!("{}", path);
-        // }
-        self.cache.iter().any(|path2| path <= path2)
-    }
-
-    pub fn update(&mut self, path: &Path) {
-        if !self.lookup(path) {
-            self.cache[self.cursor] = path.clone();
-            self.cursor = (self.cursor + 1) % self.max_size;
-        }
-    }
-}
-
-impl Default for CallInfo {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 impl Default for History {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Default for Path {
     fn default() -> Self {
         Self::new()
     }