+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support file dump in output directory.
 - Implement lookahead branching heuristic.
 
+### Removed
+
+- Remove deprecated conflict-driven heuristic.
+
 ## [0.1.3] - 2026-02-04
 
 ### Added